+++ hutool-core/src/main/java/cn/hutool/core/lang/Singleton.java
@@ -1,6 +1,7 @@
 package cn.hutool.core.lang;
 
 import cn.hutool.core.lang.func.Func0;
+import cn.hutool.core.map.MapUtil;
 import cn.hutool.core.util.ArrayUtil;
 import cn.hutool.core.util.ClassUtil;
 import cn.hutool.core.util.ReflectUtil;
@@ -52,15 +53,7 @@ public final class Singleton {
 	 */
 	@SuppressWarnings("unchecked")
 	public static <T> T get(String key, Func0<T> supplier) {
-		//return (T) POOL.computeIfAbsent(key, (k)-> supplier.callWithRuntimeException());
-		// issues#2349
-		// ConcurrentHashMap.computeIfAbsent在某些情况下会导致死循环问题，此处采用Dubbo的解决方案
-		Object value = POOL.get(key);
-		if(null == value){
-			POOL.putIfAbsent(key, supplier.callWithRuntimeException());
-			value = POOL.get(key);
-		}
-		return (T) value;
+		return (T) MapUtil.computeIfAbsent(POOL, key, (k)-> supplier.callWithRuntimeException());
 	}
 
 	/**